tests/core/test_data_message.py
#
# This file is part of the PyLECO package.
#
# Copyright (c) 2023-2023 PyLECO Developers
#
# Permission is hereby granted, free of charge, to any person obtaining a copy
# of this software and associated documentation files (the "Software"), to deal
# in the Software without restriction, including without limitation the rights
# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
# copies of the Software, and to permit persons to whom the Software is
# furnished to do so, subject to the following conditions:
#
# The above copyright notice and this permission notice shall be included in
# all copies or substantial portions of the Software.
#
# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
# THE SOFTWARE.
#

import pytest

from pyleco.core.data_message import DataMessage

topic = b"N1.CA"
cid = b"conversation_id;"
message_type = 0


@pytest.fixture
def message():
    message = DataMessage(topic, conversation_id=cid, message_type=0, data=[4, 5])
    return message


class TestDataMessageInit:
    def test_topic(self, message: DataMessage):
        assert message.topic == topic

    def test_header(self, message: DataMessage):
        assert message.header == cid + b"\x00"

    def test_data(self, message: DataMessage):
        assert message.data == [4, 5]


def test_data_message_str_topic():
    assert DataMessage(topic="topic").topic == b"topic"
